Generate image display HTML at runtime
The number of images to display depends on the number of
trace apps in the database, so the HTML that displays the plot
images can't be in a static HTML file.

Instead, read the index.html file into a string and modify it to
add the image rendering HTML before giving the HTML to the server.
This way, the static HTML file does not need to be manually
modified if the number of trace apps changes.