Handle unexpected baseline state in the database
Avoid crashing if the database does not have any baseline
records set for a given trace.  This can easily happen if a new
app is traced and the database is updated without marking any
of the records for that app as the baseline.

Fix the code to draw a baseline only if one record is marked as
a baseline for a given trace name.